fs: change offset to size_t

+++ targets/tkey/inc/fs.h
@@ -27,10 +27,10 @@ int fs_file_size(fs_file_t *f);
 int fs_read(fs_file_t *f, void *buf, size_t len);
 int fs_read_at(fs_file_t *f, void *buf, size_t len, size_t offset);
 int fs_write(fs_file_t *f, const void *buf, size_t len);
-int fs_write_at(fs_file_t *f, const void *buf, size_t len, int offset);
+int fs_write_at(fs_file_t *f, const void *buf, size_t len, size_t offset);
 
-int fs_read_open(const char *name, void *buf, size_t len, int offset);
-int fs_write_open(const char *name, const void *buf, size_t len, int offset);
+int fs_read_open(const char *name, void *buf, size_t len, size_t offset);
+int fs_write_open(const char *name, const void *buf, size_t len, size_t offset);
 
 int fs_dir_del(const char *dir_name);
 int fs_dir_nr_files(const char *dir_name);
