+++ targets/tkey/src/fs.c
@@ -120,7 +120,7 @@ int fs_init(void)
 
 // Opens and reads the file *name, if it exists, from start offset, and then
 // closes it. Returns number of bytes read or negative on error.
-int fs_read_open(const char *name, void *buf, size_t len, int offset)
+int fs_read_open(const char *name, void *buf, size_t len, size_t offset)
 {
 	lfs_file_t file;
 	struct lfs_file_config config = {0x00};
@@ -151,7 +151,7 @@ int fs_read_open(const char *name, void *buf, size_t len, int offset)
 // Opens and writes to the file *name, from start offset. Creates the file if it
 // does not exists. Closes the file once done. Returns number of bytes written
 // or negative on error.
-int fs_write_open(const char *name, const void *buf, size_t len, int offset)
+int fs_write_open(const char *name, const void *buf, size_t len, size_t offset)
 {
 
 	lfs_file_t file;
@@ -271,7 +271,7 @@ int fs_read_at(fs_file_t *f, void *buf, size_t len, size_t offset)
 
 // Writes to an already open file *f, writing to offset from *buf len
 // bytes. Returns bytes written, or negavtive on error.
-int fs_write_at(fs_file_t *f, const void *buf, size_t len, int offset)
+int fs_write_at(fs_file_t *f, const void *buf, size_t len, size_t offset)
 {
 	if (f == NULL) {
 		return -2;
